# gateway/internal/gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 多平台音乐网关 API
  version: 0.5.3
  description: |
    **多平台音乐网关 API** 是一组用于聚合 QQ 音乐 / 网易云音乐 / 酷狗音乐等多家音乐源的统一 RESTful 接口。
    网关在内部完成账号池调度、限流配额、计费扣减、Extra Usage 超额计费、余额扣款等控制逻辑，
    上游应用只需要通过统一的 API 接入即可。

    ## 协议总览
    - **基础路径**：所有业务接口位于 `/v1/...` 命名空间下。
    - **认证方式**：
      - **Cookie 会话**（前端控制台用）：登录后服务器下发 `sid` 会话 Cookie。
      - **API Key**（程序化调用用）：在 `Authorization: Bearer <api_key>` 请求头中携带。
    - **响应封装**：网关接口绝大多数返回 `{ code, message, data }` 包装结构（`code = 0` 表示成功），
      详情见下方 `Response` schema 定义。
    - **CSRF 保护**：使用 Cookie 会话时，所有非 GET/HEAD/OPTIONS 请求都需要在请求头中携带
      `X-CSRF-Token`（值与 Cookie `csrf_token` 一致）。API Key 模式不受 CSRF 限制。
    - **限流模型**：
      - **QPS**：按用户套餐限制每秒请求数。
      - **每日 / 每月配额**：免费版默认 100 次/日；超出后若开启 Extra Usage 则会按千次价从余额中扣费。

    ## 错误码
    全局错误码定义在 `ErrorCode` 字段：
    | code | 含义 |
    | ---- | ---- |
    | 0 | 成功 |
    | 400 | 请求参数错误 |
    | 401 | 未认证 / 凭据失效 |
    | 403 | 权限不足 |
    | 404 | 资源不存在 |
    | 409 | 资源冲突（重复创建、版本冲突等） |
    | 429 | 限流（QPS 或日/月配额耗尽） |
    | 500 | 服务器内部错误 |
  contact:
    name: 多平台音乐网关
    url: https://github.com/luoye663/qqmusicapi
servers:
  - url: /
    description: 当前实例
  - url: https://api.example.com
    description: 生产环境（占位，按部署实际域名替换）
...
security:
  - cookieAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/admin/users:
    get:
      tags: [用户管理（管理员）]
      summary: 列出用户
      parameters:
        - { name: email, in: query, schema: { type: string }, description: 按邮箱模糊匹配 }
        - { name: role, in: query, schema: { type: string, enum: [user, admin, super_admin] } }
        - { name: status, in: query, schema: { type: string, enum: [active, locked, disabled] } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: 用户列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserListEnvelope" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/users/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      tags: [用户管理（管理员）]
      summary: 获取用户详情
      responses:
        "200":
          description: 详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserDetailEnvelope" }
    patch:
      tags: [用户管理（管理员）]
      summary: 修改用户状态 / 角色
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [active, locked, disabled] }
                role: { type: string, enum: [user, admin, super_admin] }
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserEnvelope" }
  /v1/admin/users/{id}/password:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [用户管理（管理员）]
      summary: 重置用户密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                newPassword: { type: string, minLength: 8 }
      responses:
        "200":
          description: 已重置
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmptyEnvelope" }
  /v1/admin/users/{id}/balance:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      tags: [用户管理（管理员）]
      summary: 查询用户余额
      responses:
        "200":
          description: 余额
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BalanceEnvelope" }
    post:
      tags: [用户管理（管理员）]
      summary: 调整用户余额（增 / 减 / 覆盖）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BalanceAdjustRequest" }
      responses:
        "200":
          description: 调整后余额
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BalanceEnvelope" }
  /v1/admin/users/{id}/balance/transactions:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
      - { name: limit, in: query, schema: { type: integer, default: 50 } }
      - { name: offset, in: query, schema: { type: integer, default: 0 } }
    get:
      tags: [用户管理（管理员）]
      summary: 用户余额明细
      responses:
        "200":
          description: 明细
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BalanceTxnListEnvelope" }
  /v1/admin/users/{id}/keys:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      tags: [用户管理（管理员）]
      summary: 列出用户的 API Key
      responses:
        "200":
          description: 列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAPIKeyListEnvelope" }
  /v1/admin/users/{id}/plan:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [用户管理（管理员）]
      summary: 修改用户套餐
      description: |
        会先把 `users.plan_id` 设为新套餐，再批量更新该用户名下所有 API Key 的 plan，
        防止已有 Key 仍然走旧套餐限额。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planId]
              properties:
                planId: { type: string, example: "enterprise" }
      responses:
        "200":
          description: 修改成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSetPlanEnvelope" }

  # ========== 账号池（管理员） ==========
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: 通过 `/v1/auth/login` 颁发的会话 Cookie，浏览器自动携带；同源调用无需手动配置。
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: "在 `Authorization: Bearer <api_key>` 中携带。"
...
  responses:
...
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    Response:
      type: object
      required: [code, message]
      properties:
        code: { type: integer, example: 0, description: "0=成功；非 0 见错误码表" }
        message: { type: string, example: "ok" }
        data: { description: "实际业务数据，类型见各接口" }
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          example: { code: 400, message: "invalid body", data: null }
    EmptyEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { type: object, additionalProperties: true }
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        status: { type: string, enum: [active, locked, disabled] }
        role: { type: string, enum: [user, admin, super_admin] }
        planId: { type: string, example: "free" }
        totpEnabled: { type: boolean }
        emailVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    APIKey:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        prefix: { type: string, description: 仅返回前缀，明文 Key 只在创建时给一次 }
        planId: { type: string }
        status: { type: string, enum: [active, disabled] }
        scopes: { type: array, items: { type: string } }
        ipAllowlist: { type: array, items: { type: string } }
        rateLimitRpm: { type: integer, nullable: true }
        rateLimitDaily: { type: integer, format: int64, nullable: true }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        lastUsedAt: { type: string, format: date-time, nullable: true }
...
    Balance:
      type: object
      properties:
        userId: { type: string }
        balanceCents: { type: integer, format: int64 }
        currency: { type: string, default: CNY }
        version: { type: integer, format: int64, description: 乐观锁版本号 }
        updatedAt: { type: string, format: date-time }
    BalanceEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Balance" }
    BalanceTransaction:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        amountCents:
          type: integer
          format: int64
          description: 有符号；正为入账（充值/调整加），负为扣减（消费/调整减）
        balanceAfterCents: { type: integer, format: int64 }
        type: { type: string, enum: [topup, overage, refund, adjust_credit, adjust_debit, adjust_set] }
        description: { type: string }
        reference: { type: string, description: 关联订单/事件等外部 ID }
        createdAt: { type: string, format: date-time }
    BalanceTxnListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/BalanceTransaction" } }
                total: { type: integer }
                limit: { type: integer }
                offset: { type: integer }
...
    BalanceAdjustRequest:
      type: object
      required: [operation, amountCents]
      properties:
        operation: { type: string, enum: [credit, debit, set], description: "credit=加，debit=减，set=覆盖" }
        amountCents: { type: integer, format: int64 }
        description: { type: string }
        reference: { type: string }
...
    AdminUserEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/User" }
    AdminUserListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/User" } }
                total: { type: integer }
                limit: { type: integer }
                offset: { type: integer }
    AdminUserDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                user: { $ref: "#/components/schemas/User" }
                balance: { $ref: "#/components/schemas/Balance" }
                apiKeys: { type: array, items: { $ref: "#/components/schemas/APIKey" } }
                planCounts:
                  type: object
                  additionalProperties: { type: integer }
                  description: "key=planId，value=该套餐的 API Key 数量"
    AdminAPIKeyListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/APIKey" } }
                total: { type: integer }
    AdminSetPlanEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                planId: { type: string }
                updatedKeys: { type: integer }
                totalKeys: { type: integer }

    # ---- 账号池 ----
...